+++ Demo/build.gradle
@@ -85,9 +85,9 @@ dependencies {
     implementation project(':Card')
     implementation project(':GooglePay')
     implementation project(':LocalPayment')
-    implementation project(':PayPal')
     implementation project(':PayPalNativeCheckout')
     implementation project(':PayPalNative')
+    implementation project(':PayPal')
     implementation project(':SamsungPay')
     implementation project(':ThreeDSecure')
     implementation project(':UnionPay')
