--- PayPalNativeCheckout/src/main/java/com/braintreepayments/api/PayPalNativeCheckoutListener.java
@@ -1,8 +1,0 @@
-package com.braintreepayments.api;
-
-import androidx.annotation.NonNull;
-
-public interface PayPalNativeCheckoutListener {
-    void onPayPalSuccess(PayPalNativeCheckoutNonce nonce);
-    void onPayPalFailure(@NonNull Exception error);
-}
